src/ctkserver/requirements.txt
msgpack-python==0.4.8
PyMySQL==0.7.11
SQLAlchemy==1.2.0b2
